Remove Settings and refactor Solver4x4
