+++ src/berreman4x4/solver2x2.py
@@ -4,6 +4,8 @@ from numpy.lib.scimath import arcsin, sqrt
 
 from .solver import Solver
 from .math import unitConversion
+from .result import Result
+
 
 class Solver2x2(Solver):
     '''
@@ -54,10 +56,10 @@ class Solver2x2(Solver):
     def list_snell(self, n_list):
         angles = arcsin(n_list[0] * np.sin(np.deg2rad(self.theta_i)) / n_list)
 
-        angles[0] = np.where(np.invert(Solver2x2.is_forward_angle(n_list[0], angles[0])), 
-                            np.pi - angles[0], angles[0])
-        angles[-1] = np.where(np.invert(Solver2x2.is_forward_angle(n_list[-1], angles[-1])), 
-                            np.pi - angles[-1], angles[-1])
+        angles[0] = np.where(np.invert(Solver2x2.is_forward_angle(n_list[0], angles[0])),
+                             np.pi - angles[0], angles[0])
+        angles[-1] = np.where(np.invert(Solver2x2.is_forward_angle(n_list[-1], angles[-1])),
+                              np.pi - angles[-1], angles[-1])
 
         return angles
 
@@ -80,7 +82,7 @@ class Solver2x2(Solver):
         th_list = self.list_snell(n_list)
         kz_list = 2 * np.pi * n_list * np.cos(th_list) / lbda
 
-        delta = kz_list[1:-1] * (d_list if n_list.ndim == 1 else d_list[:,None])
+        delta = kz_list[1:-1] * (d_list if n_list.ndim == 1 else d_list[:, None])
 
         esum = 'ij...,jk...->ik...'
         ones = np.repeat(1, n_list.shape[1]) if n_list.ndim > 1 else 1
@@ -91,23 +93,24 @@ class Solver2x2(Solver):
         Mp = np.array([[ones, rp],
                        [rp, ones]], dtype=complex) / tp
 
-        for i in range(1, num_layers-1):
-            rs, rp, ts, tp = Solver2x2.fresnel(n_list[i], n_list[i+1], th_list[i], th_list[i+1])
-            em = np.exp(-1j * delta[i-1])
-            ep = np.exp(1j * delta[i-1])
+        for i in range(1, num_layers - 1):
+            rs, rp, ts, tp = Solver2x2.fresnel(n_list[i], n_list[i + 1], th_list[i], th_list[i + 1])
+            em = np.exp(-1j * delta[i - 1])
+            ep = np.exp(1j * delta[i - 1])
 
             Ms = np.einsum(esum, Ms,
                            np.array([[em, rs * em],
                                      [rs * ep, ep]], dtype=complex) / ts)
             Mp = np.einsum(esum, Mp,
-                            np.array([[em, rp * em],
-                                      [rp * ep, ep]], dtype=complex) / tp)
+                           np.array([[em, rp * em],
+                                     [rp * ep, ep]], dtype=complex) / tp)
 
-        self._rtots = Ms[1,0] / Ms[0,0]
-        self._ttots = 1 / Ms[0,0]
-        self._rtotp = Mp[1,0] / Mp[0,0]
-        self._ttotp = 1 / Mp[0,0]
+        self._rtots = Ms[1, 0] / Ms[0, 0]
+        self._ttots = 1 / Ms[0, 0]
+        self._rtotp = Mp[1, 0] / Mp[0, 0]
+        self._ttotp = 1 / Mp[0, 0]
 
+        return Result(self)
 
     @staticmethod
     def fresnel(n_i, n_t, th_i, th_t):
@@ -135,8 +138,8 @@ class Solver2x2(Solver):
         return r_s, r_p, t_s, t_p
 
     @staticmethod
-    def is_forward_angle(n, theta): 
+    def is_forward_angle(n, theta):
         ncostheta = n * np.cos(theta)
         return np.where(abs(ncostheta.imag) > 1e-10,
-                        ncostheta.imag > 0, 
+                        ncostheta.imag > 0,
                         ncostheta.real > 0)
